Initial implementation of the ExecutionContext.
I created two new folders in framework: api and context. The api folder will be used to contain interfaces for commonly used framework artifacts. The implementations will remain in the existing components. The context folder contains the implementation of ExecutionContext.

I was able to solve the cross-dependency issue by having each component in the compiler chain extend the ExecutionContext interface and add its own artifacts. The final version of the interface is in the service component. I know this is unusual, but it is only temporary. When the api folder is built out, the ExecutionContext interfaces will be consolidated into one - and that will be kept in the api folder.

Right now the ExecutionContext doesn't do anything except contain commonly used objects, and it keeps track of the thread's execution path. The idea is to have this one object passed around instead of the current half-dozen or so objects. The execution path will be used for the forthcoming security redesign.

If you want to see it in action, change the executionContext.verbose property in api.properties to true, and uncomment the blocks of screen widget code in the Example component: line 40 in ExampleScreens.xml, and line 25 in ExampleForms.xml. The console log will show the ExecutionContext pushing and popping artifacts, and the Find Example screen will display the execution path for the screen.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/executioncontext20090812@803693 13f79535-47bb-0310-9956-ffa450edef68