"Applied fix from trunk for revision: 883933" ------------------------------------------------------------------------ r883933 | doogie | 2009-11-25 01:31:25 +0100 (mer. 25 nov. 2009) | 5 lignes
In the Product/Parties screen, the link to the related party was broken.
Instead of showing the related party, it would show the user who was
logged into the partymgr backend.  The patch that fixed this was
provided by Ean Schuessler in
(https://issues.apache.org/jira/browse/OFBIZ-3258) OFBIZ-3258
------------------------------------------------------------------------

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@884020 13f79535-47bb-0310-9956-ffa450edef68