Migrated changes from trunk r946100
Removed an unnecessary decorator screen CommonQuotePriceDecorator related to Order Manager Quotes. The only difference between this and screen 
CommonQuoteDecorator is that the the former has a permission check for ORDERMGR__QUOTE_PRICE where as latter has a check for permission 
ORDERMGR_VIEW. The permission ORDERMGR_QUOTE_PRICE is though already checked at menu level and hence is not required explicity
at the decorator level.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches@946948 13f79535-47bb-0310-9956-ffa450edef68