"Applied fix from trunk for revision: 1002963" ------------------------------------------------------------------------ r1002963 | jleroux | 2010-09-30 10:49:15 +0200 (jeu. 30 sept. 2010) | 16 lignes
A modified patch from Hans Bakker "Ecommerce Customer Request has error" (https://issues.apache.org/jira/browse/OFBIZ-3965) - OFBIZ-3965

BJ: Login from ecomerce as DemoCustomer
click on
Create Customer Request
put in product # WG-1111
click on Add to cart
Following Errors Occurred:

Customer Request should have a 'Name' if not in the draft status

Hans: We decided that a customer request needs a name of the request when it is not in the draft status.
I try to find the 'create customer request' option in e-commerce but i cannot find it.
Attached a diff file which could solve the problem, but could not test it. of course need internationalisation.

JLR: I provided the i18n, but weirdly it always render in English (I tested in French and double-checked the labels I created)
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1002968 13f79535-47bb-0310-9956-ffa450edef68