"Applied fix from trunk for revision: 1042348" ------------------------------------------------------------------------ r1042348 | jleroux | 2010-12-05 14:37:07 +0100 (dim., 05 déc. 2010) | 15 lines
A patch from Anne Jessel "Collapse all broken if hyperlink in pane" (https://issues.apache.org/jira/browse/OFBIZ-4022) - OFBIZ-4022

To reproduce the bug:
    * edit the EditProduct form in applications/product/widget/catalog/ProductForms.xml
    * add a new field with a hyperlink sub-element.
    * refer to this new field in the first, non-collapsible field-group in the sort-order
    * go to this page in the UI (edit a product in the catalog)
    * the Expand All and Collapse All buttons do not work properly

The problem is caused by the javascript finding the added hyperlink field in the first field-group, and incorrectly assuming it is
the expand/collapse link for the field-group. This causes the javascript to crash when it gets a null reference to an enclosing li element.

It is not specific to the EditProduct form, but applies to any group of collapsible panes where one contains a hyperlink.

The attached patch simply checks that the li element actually exists.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1042358 13f79535-47bb-0310-9956-ffa450edef68