Applied fix from trunk for revision: 1043998 ===
When thte findByAnd->findList change was done, the parameters to the new
findList method were added, but the actual method name itself was not
changed.  However, since there was a vararg version of
findByAnd(Object...), this meant the code actually compiled anyways, so
the problem was not detected.

broken since: 654934
date: Fri May 9 20:57:23 2008 +0000


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1044074 13f79535-47bb-0310-9956-ffa450edef68