"Applied fix from trunk for revision: 1094126" ------------------------------------------------------------------------ r1094126 | jleroux | 2011-04-17 12:15:23 +0200 (dim., 17 avr. 2011) | 8 lines
A patch from Pierre Gaudin "Error in stock reception on same inventory item with same unit price" https://issues.apache.org/jira/browse/OFBIZ-4250

During the reception of stock, if the reception is made on a line of stock which already exists with the same unit price, an accounting transaction due to the stock price value change is created with an amount of debit and credit of 0.

This is not normal and generate an error in the service postAcctgTrans.

This aim of this patch is to avoid creation of accounting transaction linked to the stock price value change if the unit price does not change.

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release09.04@1094131 13f79535-47bb-0310-9956-ffa450edef68