"Applied fix from trunk for revision: 1167116  " ------------------------------------------------------------------------ r1167116 | jleroux | 2011-09-09 13:50:01 +0200 (ven., 09 sept. 2011) | 3 lines
A patch from Paul Foxworthy "roleTypeId not copied from ProductStoreShipmentMethod to OrderItemShipGroup" https://issues.apache.org/jira/browse/OFBIZ-4389

In ordershippinginfo.ftl, a list of shipping methods is presented, and when one is chosen values are written into an OrderItemShipGroup. The shipping methods are defined by rows in ProductStoreShipMethod, and one of the attributes of this entity is a carrierRoleTypeId. Similarly, OrderItemShipGroup has a carrierRoleTypeId, but no data is being written into it.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1167136 13f79535-47bb-0310-9956-ffa450edef68