Partially "Applied fix from R12.04 for revision: 1493353"  PortalPageWorker.java not concerned ------------------------------------------------------------------------ r1493352 | jleroux | 2013-06-15 15:40:42 +0200 (sam., 15 juin 2013) | 40 lines
A patch from Olivier Heintz for "Portal page personalization is not working" https://issues.apache.org/jira/browse/OFBIZ-5189

A bug in portal page personalization.
When trying to delete portlet or column in portal page, there is an error because in delete service, cache is used to read data before remove.

Correction proposed is to remove use-cache="true"
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release10.04@1493356 13f79535-47bb-0310-9956-ffa450edef68