"Applied fix from trunk for revision: 1509164" ------------------------------------------------------------------------ r1509164 | jleroux | 2013-08-01 12:47:02 +0200 (jeu. 01 août 2013) | 3 lignes
A patch from János Zsolt for "Load cart from order loses OrderItemAttributes" https://issues.apache.org/jira/browse/OFBIZ-5286

When an order is created from a cart, OrderItemAttribute entities are created as expected. But when this same order is used to load back to the cart, the OrderItemAttributes are not reset. I found the cause of this issue in the org.ofbiz.order.shoppingcart.ShoppingCartServices.loadCartFromOrder(DispatchContext, Map<String, Object>) method.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1533563 13f79535-47bb-0310-9956-ffa450edef68