Applied fix from trunk for revision: 1554373 ===
Removed incomplete JUnit test that was created by Hans Bakker in commit 1449615 in order to test the feature associated to a new field calcTaxAfterPromotion; after this, Paul Foxworthy reverted the commit in rev. 1468865 but he forgot to remove the whole test and instead he updated it, leaving it in a useless state.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1554374 13f79535-47bb-0310-9956-ffa450edef68