"Applied fix from trunk for revision: 1560699" ------------------------------------------------------------------------ r1560699 | jleroux | 2014-01-23 15:35:20 +0100 (jeu. 23 janv. 2014) | 5 lignes
A patch from Justen Walker for "Missing import in FindGeneric.groovy" https://issues.apache.org/jira/browse/OFBIZ-5504

Revision 1560176 introduces a change to FindGeneric.groovy that extracts some logic into UtilMisc. However, this class is not imported in the file and results in a MissingProperty exception.
This patch adds the import statement which resolves this bug.

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release11.04@1562107 13f79535-47bb-0310-9956-ffa450edef68