"Applied fix from trunk for revision: 1571219" ------------------------------------------------------------------------ r1571219 | jleroux | 2014-02-24 11:33:12 +0100 (lun. 24 févr. 2014) | 10 lignes
After reverting r1571198 (there was too much unwanted changes due to my "Save Actions" setting in Eclipse) doing it again with the only changes really wanted

A slightly modified patch from Praveen Agrawal for "The getFeatureIdQtyMap function doesn't include the DISTINGUISHING_FEAT into the featureMap" https://issues.apache.org/jira/browse/OFBIZ-5544

The getFeatureIdQtyMap function in ShoppingCartItem.java doesn't include the DISTINGUISHING_FEAT in featureMap so if we want to calculate the Shipping amount based on Distinguishing feature, this method doesn't return the Distinguishing Feature in Feature Map. If we add DISTINGUISHING_FEAT as productFeatureApplTypeId in filterExprs then the feature map would contain the Distinguishing feature in addition to Standard and Required Feature.

Also in ShoppingCartItem.java, putAdditionalProductFeatureAndAppl method, i think the OrderAdjustment should only be included when the amount is not null.

jleroux: When the amount is null it's now not added in the adjustement (I moved that in the block above where the amount is already tested for null). But the OrderAdjustment is still created because a recurringAmount could still exist . I just added a warning for when both amounts are null. We could decide to not create an adjustement when there are no amounts at all. But I preferred a warning, because it's a weird situation and people could prefer to be aware...

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1571226 13f79535-47bb-0310-9956-ffa450edef68