Applied fix from trunk for revision: 1570639 ===
Upgrade to Tomcat 7.0.52 (latest bug fix release). I have used the standard release and not the "embedded" version after I have realized (from a conversation with Mark Thomas @ Tomcat and then directly checking the content of each jar) that the embedded jars contain also the J2ee apis that we already have in the OFBiz classpath.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1571255 13f79535-47bb-0310-9956-ffa450edef68