Merged revision(s) 1617936 from ofbiz/trunk: Fixed a bug in Start.java reported by Justen Walker (https://issues.apache.org/jira/browse/OFBIZ-5710):
The singleton pattern was broken by adding Commons Daemon support code to the class - resulting in multiple instances when using Commons Daemon.

I extracted the Commons Daemon code to a separate class - which delegates to the singleton Start instance. Users of Commons Daemon will need to update their scripts/etc to point to the new class instead of Start.

I also added a comment describing the design of Start.java, so hopefully this will help others maintain the code without breaking it.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1617951 13f79535-47bb-0310-9956-ffa450edef68