"Applied fix from trunk for revision: 1650642  " ------------------------------------------------------------------------ r1650642 | jleroux | 2015-01-09 21:19:17 +0100 (ven. 09 janv. 2015) | 21 lignes
A patch from Wai for "state/province dropdown contains incorrect items when Canada is chosen in country dropdown" https://issues.apache.org/jira/browse/OFBIZ-5933



When creating a postal address, selecting Canada for country automatically filled the state/province dropdown with values. The issue is that the values of the state/province dropdown list are incorrect.

I was expecting the options for the state/province dropdown to have the following:
name=Alberta, value=AB
name=British Columbia, value=BC
...

What is actually present in the drop down is:
name=AB, value=AB
name=AB, value=BC
...

One observation is that this issue does not appear when the country selected is the United States. It seems to happen with Canada.
Another observation is that this issue applies to the 'en' locale. It seems to work fine for 'fr' locale.

jleroux: this was OK for Saskatchewan, that leaded to the solution

------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1650645 13f79535-47bb-0310-9956-ffa450edef68