Applied fix from trunk for revision: 1665162 ===
Fix for issue happening when a service event was executed using JSON (i.e. parameters in the request body) after a getParameter was called (by other framework code) to fetch the parameter from the body; in that case an io error was thrown because the reuest body stream can be read only once.
This fix move the code that reads the JSON parameters from the body into the ContextFilter: the parameters are then stored as attributes that can be later used by events and other artifacts (service events, simple events, scripts events support this and Java events if they use UtilHttp.getCombinedMap(request) to get the map of input parameters. 



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1665163 13f79535-47bb-0310-9956-ffa450edef68