A patch from Martin Decker for "Release branches fail to compile with JDK 1.8" https://issues.apache.org/jira/browse/OFBIZ-6252
Due to a bug that was already fixed in trunk by Adrian C. (Jan 2015) but not in the current release branches (I checked 12.04, 13.07 and 14.12), the branches  could not be compiled using JDK 1.8 because it discovers an "variable operator might not have been initialized" error in ModelViewController.java.

Furthermore this is also a logical bug that leads to an incomplete error message.

So the trivial fix should be applied to the release branches, I will attach one for 14.12.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1674364 13f79535-47bb-0310-9956-ffa450edef68