Applied fix from trunk for revision: 1674908 ===
OFBIZ-6269: "Pro-rate shipping calculation is generating incorrect invoice in case of order with subtotal $0.00"
The solution provided in this commit is to use, in these cases, the quantities instead of the amounts for pro rating of misc and shipping adjustments.
Thanks to Mohammad Kathawala for the bug report, great analysis and tests that have helped me a lot to implement this fix.




git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1674925 13f79535-47bb-0310-9956-ffa450edef68