Applied fix from trunk for revision: 1697590 ===
Reverted 1530237: the sql type for Postgres for "comments" and "description" fields was changed from VARCHAR(255) to TEXT; the new TEXT type allows strings of any length; however all other types definitions (MySQL, Derby etc...) are using VARCHAR(255) thus limiting the length of "comments" and "description" fields to 255 characters.
By reverting this change we make sure that the Postgres db will be consistent with the other dbs.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1697593 13f79535-47bb-0310-9956-ffa450edef68