"Applied fix from trunk for revision: 1704052" ------------------------------------------------------------------------ r1704052 | mbrohl | 2015-09-19 18:02:15 +0200 (sam. 19 sept. 2015) | 5 lignes
Applied patch for OFBIZ-6635: Old UserLogin from userLoginId-change is not correctly disabled.

If a userLoginId of an existing user is updated by LoginServices.updateUserLoginId, a new UserLogin value is created with the data of the old one and the old one is disabled afterwards. In addition to switch the enabled flag to "N" the disabledDateTime is set to current date. This is wrong because this makes it possible to reenable the old UserLogin by just do a login with the old userLoginId (standard mechanism to lock the login for a while after subsequent failed login requests).

Thanks Martin Becker for spotting this and providing the patch.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release12.04@1704067 13f79535-47bb-0310-9956-ffa450edef68