Applied fix from trunk for revision: 1706577 ===
(OFBIZ-6212) Fixed accounting report PDF export issue, it was throwing FOP exception. It was due to empty description string, if description is empty then makeHyperlinkString does not call the makeHyperlinkString macro to render the hyper link and due to this fo:table-cell is render without fo:block and its not a valid syntax of fo:table-cell.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1706583 13f79535-47bb-0310-9956-ffa450edef68