"Applied fix from trunk for revision: 1736851  " ------------------------------------------------------------------------ r1736851 | jleroux | 2016-03-28 11:55:17 +0200 (lun. 28 mars 2016) | 25 lignes
A patch from James Yong for <<popup fades in and out non-stop when hovers on a product image in list view>> https://issues.apache.org/jira/browse/OFBIZ-6961

Problem
========
Go to https://localhost:8443/ecommerce/gift-cards-GC-100-c
When the mouse hovers on a product image in a list view, the popup fades in and out non-stop.
 
Cause
=====
The list view has a narrower height than the grid (matrix) view. This causes the popup to overlay on the target and results in the cycling.

Quick Fix
==============
To illustrate the problem, we change the display to grid view, instead of list view, by append the following to the CatalogDetail.groovy file
{quote}
context.put("numCol", 3);
context.viewSize = 12;
{quote}

Better Solution
===============
Use tooltip from jQuery UI. Works correctly when the products are displayed in list view / grid view. A patch will be provided.


jleroux: I checked, there should not be other cases like this
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release15.12@1736855 13f79535-47bb-0310-9956-ffa450edef68