"Applied fix from trunk for revision: 1743230" ------------------------------------------------------------------------ r1743230 | jleroux | 2016-05-10 19:12:05 +0200 (mar. 10 mai 2016) | 3 lignes
A patch from Ingo Wolfmayr for "Update empty blog article text does not work" https://issues.apache.org/jira/browse/OFBIZ-7062

When creating a blog-entry with empty article text, it is not possible to add update the article. Content entry is created but not assigned to contentId of blog-entry. The attached patch fixes this.
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1743233 13f79535-47bb-0310-9956-ffa450edef68