Manually applied fix from trunk revision 1747202. ---------------------------------------------------------------------
[OFBIZ-7209] Fixed Order Ship Group PDF.
This issue was due to the declaration of map with list syntax. It was introduced when we moved from Javolution collections to java collections.
One more issue found after fixing map syntax, that is missing block elements error. Added checks to handle that error as well.

Thanks Amardeep Singh Jhajj for the contribution.
---------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release15.12@1747203 13f79535-47bb-0310-9956-ffa450edef68