Backports r1765094 (editing by hand) ======================================================================================== Fixed: You cannot edit an image in content (OFBIZ-7000)
If you get to
content/control/EditLayout?drDataResourceId=NOCONTENTFOUND&contentId=NOCONTENTFOUND
When trying to editing you get an error

Mohammed Rehan Khan explored the issue and found that there is a condition in 
DataResourceWorker method to render content of 'text' mime type only, for others
 it returns exception.

We can fix this by setting up mimeTypeID as null in request param

Question: Why setting up mimeTypeTypeId as 'null' is fixing this issue?
Answer: As we don't have image specific worker method for rendering these type 
of contents, so we are using DataResourceWorker.renderDataResourceAsText() 
to render image type content as well. Due to following condition, 
targetMimeTypeId is setting up as "text/html" if sent null and hence not 
throwing the exception reported in the ticket.

Thanks: Mohammed Rehan Khan for the patch
========================================================================================

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release14.12@1765096 13f79535-47bb-0310-9956-ffa450edef68