"Applied fix from trunk framework for revision: 1784549  " ------------------------------------------------------------------------ r1784549 | jleroux | 2017-02-27 13:33:42 +0100 (lun. 27 févr. 2017) | 26 lignes
Fixed: "Login and logout process in demos shows a certificate issue"
(OFBIZ-9206)

Also fixes "16.11 ofbizUrl include host+port and break some reverse-proxy/docker
 setups"
(OFBIZ-9224)

It was an easy fix, I just imported
<SystemProperty systemPropertyId="port.https" systemResourceId="url" 
systemPropertyValue=""/>
in trunk demo and all work perfectly.

I also tried to replace locally
port.https=8443
by
port.https=
in url.properties (w/o SystemProperty) and did not face any issue but with 
portOffset. This is due to the WebSiteProperties class works and there is also 
an easy fix: don't add twice the portOffset when it's build from the request, 
and only then. Keep it as is when it's build from a WebSite GenericValue. 
We then trust the user and don't rely on the request.

I also removed the deprecated RequestHandler.getDefaultServerRootUrl() 
I think it was time...

Thanks: Pierre Smits and Leonard Lin for reports
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1784550 13f79535-47bb-0310-9956-ffa450edef68