Fixed: "Login and logout process in demos shows a certificate issue" (OFBIZ-9206)
After my conclusions at OFBIZ-9240 I decided to reapply the changes

Here, unlike in trunk and stable, I kept getDefaultServerRootUrl() 
and did no changes in url.properties. 
URL properties will be, for instance, 
used to specify the ports used in demos (28443/28080)


The idea is to use the port initially passed by a successful login request as 
it's done since WebSiteProperties class exists. I also don't add twice the 
portOffset when it's build  from the request, and only then. 
Keep it as is when it's build from a WebSite GenericValue. 
We then trust the user and don't rely on the request.

Thanks: Pierre Smits for report

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1785098 13f79535-47bb-0310-9956-ffa450edef68