"Applied fix from trunk framework for revision: 1789863  " ------------------------------------------------------------------------ r1789863 | jleroux | 2017-04-02 11:26:36 +0200 (dim. 02 avr. 2017) | 33 lignes
Fixed: Sort Links in Lookup for Data Resource Id causes unwanted behaviour on 
Find Content page
(OFBIZ-9280)

How to reproduce :
1. Log in the Content component
2. Go to the Content tab. https://localhost:8443/content/control/findContent
3. In the search form, open the lookup of the field ' Data Resource Id '
4. Click on any of the table header links to sort the table. Results will be
 displayed on a new unstyled window closing the Lookup dialog.


Problem:
Problem:
1. Lookup dialog for dataResourceId uses "ListLookupDataResource" form to list 
Data Resource records.
2. "ListLookupDataResource" form extends "ListDataResource" form for all the 
fields.
3. Sort-field is set to true in "ListDataResource" form for all the fields.
4. Sort-field adds <a> hyperlink to "LookupResource" .
5. When clicked <a> link it moves to a new page and doesn't renders the response
 in the Lookup dialog.
Solution:
First of all it is a lookup to select DataResourceId so as the pattern followed 
in other lookups in OFBiz it should not have any sort-fields.To set it false, 
it should be overridden. If all fields needs to be overridden it is better not 
to extend "ListDataResource" form and add these fields right away.
Apart from that added 'widget-style="smallSubmit"' to dataResourceId according 
to the uniform pattern followed.

jleroux: I added the header-row-style="header-row-2" style

Thanks: Aditya Sharma
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1789864 13f79535-47bb-0310-9956-ffa450edef68