"Applied fix from trunk framework for revision: 1789874  " ------------------------------------------------------------------------ r1789874 | jleroux | 2017-04-02 14:11:44 +0200 (dim. 02 avr. 2017) | 29 lignes
Fixed: Sort Links in Lookup for Owner Content Id causes unwanted behaviour on 
Find Content page
(OFBIZ-9279)

How to reproduce :
1. Log in the Content component
2. Go to the Content tab. https://localhost:8443/content/control/findContent
3. In the search form, open the lookup of the field ' Owner Content Id '
4. Click on any of the table header links to sort the table. Results will be 
displayed on a new unstyled window closing the Lookup dialog.

Problem:
1. Lookup dialog for ownerContentId uses "ListLookupContent" form to list 
   Content records.
2. "ListLookupContent" form extends "ListContent" form for all the fields.
3. Sort-field is set to true in "ListContent" form for all the fields.
4. Sort-field adds <a> hyperlink to "LookupContent" .
5. When clicked <a> link it moves to a new page and doesn't renders the response
   in the Lookup dialog.

Solution:
First of all it is a lookup to select ownerContentId so as the pattern followed
 in other lookups in OFBiz it should not have any sort-fields. 
To set it false, it should be overridden. If all fields needs to be overridden 
it is better not to extend "ListContent" form and add these fields right away.
Apart from that added 'header-row-style="header-row-2"' to form.


Thanks: Aditya Sharma
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1789875 13f79535-47bb-0310-9956-ffa450edef68