Applied fix from trunk for revision: 1799417 ===
Fixed: unit test failure when "verbose" logging was on.
(OFBIZ-9305)

When print.verbose was set to true in debug.properties, the unit test 
EntitySaxReaderTests.parse was failing because it is using some mock objects; 
fixed by temporarily disabling the verbose logging during the test.

Thanks: William Wang for the report.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1799419 13f79535-47bb-0310-9956-ffa450edef68