"Applied fix from trunk framework for revision: 1801094" ------------------------------------------------------------------------ r1801094 | jleroux | 2017-07-06 21:02:01 +0200 (jeu. 06 juil. 2017) | 16 lignes
Fixed: Missing secure port in URLs when sending mails
(OFBIZ-9449)
While working on an issue reported by Pierre Gaudin on user ML 
https://s.apache.org/gDo3, here quoted:

I found another issue while testing: the port 8443 is missing  in the URL used 
underneath by the form reported by Pierre.

It was broken by OFBIZ-9206 where I removed the value of port.https in 
url.properties. 

It's not an issue if you set the value of port.https, or let it empty when 
using a SSL certificate on a production server. But if you try locally 
(localhost) then the URL misses the 8443 port.

Solution append 8443 as default secure port when using localhost
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1801097 13f79535-47bb-0310-9956-ffa450edef68