Manually applied fix from trunk for revision: 1802690 ============================
Fixed: Adding condition in price rule on 'Role Type' showing incorrect result.
(OFBIZ-7275)

The issue exists because in role type list there are multiple occurrences of substring "CUSTOMER". The existing check was just checking the occurrence of a substring in the string and found this value more than once in the list so the last option gets selected. Applied equality check to find the exact match.

Thanks Aditi Patidar for reporting and providing the patch.

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1802692 13f79535-47bb-0310-9956-ffa450edef68