Applied fix from trunk for revision: 1802766 ===
Fixed: MRP generates duplicate requirements when items are initially under 
minimum quantity and have with pending approved purchase orders.
(OFBIZ-9526)

With this change the requirements to replenish the inventory up to the minimum 
stock quantity is generated, if needed, only when the first demand event is 
processed or at the end.

Thanks: Amit Gadaley for the help in recreating the issue and testing the fix.



git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1802773 13f79535-47bb-0310-9956-ffa450edef68