"Applied fix from trunk framework for revision: 1804843" ------------------------------------------------------------------------ r1804843 | jleroux | 2017-08-11 21:14:21 +0200 (ven. 11 août 2017) | 12 lignes
Fixed: Missing action in empty if-statement in accounting.PaymentGatewayServices
(OFBIZ-9530)

In the method processAuthResult() line 1889 there is an empty if-statement for 
the case that authResult is null. It seems like somebody was planning to cover 
the case but forgot to add code.

I would suggest to return a service error when the authResult is null since 
otherwise it cannot be determined if the payment is authorized or not.


Thanks: Kyra Pritzel-Hentley
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1804845 13f79535-47bb-0310-9956-ffa450edef68