"Applied fix BY HAND from trunk framework for revision: 1806266" ------------------------------------------------------------------------ r1806266 | jleroux | 2017-08-26 09:37:49 +0200 (sam., 26 août 2017) | 23 lines
Fixed: Improvements in LabelReferences class
(OFBIZ-9623)

For Java:
Current support is only for occurrences like "UtilProperties.getMessage(",
 but there are other occurrences such as "uiLabelMap.get(" 
in Java as well. Need to add support for same.

For groovy:
After some recent improvements for groovy handling, still, there is a 
possibility that someone is using "UtilProperties.getMessage(" in groovy files 
as well, so IMO, that must also be handled.

For XML forms:
In XML forms current support in handled for auto-fields-service and 
auto-fields-entity but there are some occurrences of tooltip element of 
field tag as well, that must also be handled.

For XML Simple Methods:
Doesn't handle case of property-to-field and default-message elements. 
(Separate ticket created OFBIZ-9606, just adding it here for reference)

Thanks: Suraj Khurana
------------------------------------------------------------------------

git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release13.07@1806271 13f79535-47bb-0310-9956-ffa450edef68