"Applied fix from trunk framework for revision: 1814349  " ------------------------------------------------------------------------ r1814349 | jleroux | 2017-11-05 12:28:42 +0100 (dim., 05 nov. 2017) | 12 lines
Fixed: Secure the login.secret_key_string
(OFBIZ-9966)

When OFBIZ-4983 was implemented I missed that we put the login.secret_key_string
 as a property in security properties. This should not have been because it 
eases attackers work.

The recommended way is to have it as a private static final String that can be 
changed just when compiling using sed and uuidgen. So then the key is temporay 
and final and it gets quite harder for a possible attacker to use this mean.


------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1814352 13f79535-47bb-0310-9956-ffa450edef68