Applied fix from trunk for revision: 1819136 ===
Fixed: Party Classification Groups are not listing on FindParty screen, applied patch from jira issue (OFBIZ-10007)
Fixed issue in which classification is not listed in the dropdown.
First value is undefined in the list, this is because of asmSelect Jquery which will cover in another ticket.
Thanks, Aman Agrawal and Sonal Patwari for your contribution.


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1819137 13f79535-47bb-0310-9956-ffa450edef68