"Applied fix from trunk framework for revision: 1821613  " ------------------------------------------------------------------------ r1821613 | jleroux | 2018-01-19 11:33:05 +0100 (ven., 19 janv. 2018) | 15 lines
Fixed: Not able to add the requirement,broken screen with the error message in 
workeffort component
(OFBIZ-10122)

Steps to regenerate:
1) Open URL https://demo-trunk.ofbiz.apache.org/sfa/control/Events
2) Click on the event created.
3) To add the requirements of events, click on the "Requirements" button.
4) Add the existing requirement Id or add a new requirement. Click on Save.

Actual: Not able to add the requirement. Broken screen with the error message.

Note: Refer the screenshot attached. 

Thanks: Rubia Elza Joshep for report and Yogesh Naroliya for the patch
------------------------------------------------------------------------


git-svn-id: https://svn.apache.org/repos/asf/ofbiz/branches/release16.11@1821616 13f79535-47bb-0310-9956-ffa450edef68